Changed 'Day' to 'Today'  in .events-nav
Changed to facilitate the removal of the 'Return to today' link. There is concern for how much event content is visible at mobile widths. This link is redundant (there are 2 other ways to access today's events—the date on the calendar widget and the above 'Day'/'Today' link) and it pushes content down.