Remove 'Return to Today' button/link
There is concern for how much event content is visible at mobile widths. This link is redundant (there are 2 other ways to access today's events—the date on the calendar widget and the above 'Day'/'Today' link) and it pushes content further down the screen.

I think in future sprints we should look at ways to incorporate this link's AJAX into other areas of Events.